+++ src/DamYou/Platforms/Windows/App.xaml.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.UI.Xaml;
+using DamYou.Services;
+using Serilog;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.